Move handleScrollToCommentOnLoad() helper to different module
