+++ packages/liveblocks-react/src/room.tsx
@@ -87,11 +87,11 @@ import type {
   RoomNotificationSettingsStateSuccess,
   RoomProviderProps,
   ThreadsState,
-  ThreadsStateResolved,
   ThreadsStateSuccess,
   ThreadSubscription,
   UseThreadsOptions,
 } from "./types";
+import { handleScrollToCommentOnLoad } from "./use-scroll-to-comment-on-load-effect";
 
 const noop = () => {};
 const identity: <T>(x: T) => T = (x) => x;
@@ -744,43 +744,6 @@ function makeRoomContextBundle<
     }
 
     throw innerError;
-  }
-
-  /**
-   * Scroll to the comment with the ID in the hash of the URL based on whether
-   * the query is loading and whether the hook should scroll to the comment on load.
-   */
-  function handleScrollToCommentOnLoad(
-    isQueryLoading: boolean,
-    shouldScrollOnLoad: boolean,
-    state: ThreadsStateResolved<TThreadMetadata>
-  ) {
-    if (shouldScrollOnLoad === false) return;
-
-    if (isQueryLoading === true) return;
-
-    const isWindowDefined = typeof window !== "undefined";
-    if (!isWindowDefined) return;
-
-    const hash = window.location.hash;
-    const commentId = hash.slice(1);
-
-    // If the hash is not a comment ID, we do not scroll to it
-    if (!commentId.startsWith("cm_")) return;
-
-    // If a comment with the ID does not exist in the DOM, we do not scroll to it
-    const comment = document.getElementById(commentId);
-    if (comment === null) return;
-
-    const comments = state.threads.flatMap((thread) => thread.comments);
-    const isCommentInThreads = comments.some(
-      (comment) => comment.id === commentId
-    );
-
-    // If the comment is not in the threads for this hook, we do not scroll to it
-    if (!isCommentInThreads) return;
-
-    comment.scrollIntoView();
   }
 
   function useThreads(
